Add files via upload
5V SMD REGULATOR
NCV4274, NCV4274A

Regulator Family,

400 mA, 2% and 4% Voltage

 

Description The NCV4274 and NCV4274A is a precision micro−power voltage regulator with an output current capability of 400 mA available in the DPAK, D2PAK and SOT−223 packages. The output voltage is accurate within ±2.0% or ±4.0% depending on the version with a maximum dropout voltage of 0.5 V with an input up to 40 V. Low quiescent current is a feature drawing only 150 A with a 1 mA load. This part is ideal for automotive and all battery operated microprocessor equipment. The regulator is protected against reverse battery, short circuit, and thermal overload conditions. The device can withstand load dump transients making it suitable for use in automotive environments.



Features

• 2.5, 3.3 V, 5.0 V, 8.5 V, ±2.0% Output Options

• 2.5, 3.3 V, 5.0 V, ±4.0% Output Options

• Low 150 A Quiescent Current at 1 mA load current

• 400 mA Output Current Capability

• Fault Protection

• +60 V Peak Transient Voltage with Respect to GND −42 V Reverse Voltage Short Circuit Thermal Overload

• Very Low Dropout Voltage

• NCV Prefix for Automotive and Other Applications Requiring Unique Site and Control Change Requirements; AEC−Q100 Qualified and PPAP Capable

• These are Pb−Free Devices